dashboard/requirements.txt
streamlit
google-cloud-storage
Pillow
python-dotenv
